Commit clean library files into working tree and clean kernel source

+++ projects/util_libs/libplatsupport/src/plat/pc99/acpi/browser.c
@@ -1,13 +1,11 @@
 /*
- * Copyright 2017, Data61
- * Commonwealth Scientific and Industrial Research Organisation (CSIRO)
- * ABN 41 687 119 230.
+ * Copyright 2014, NICTA
  *
  * This software may be distributed and modified according to the terms of
  * the BSD 2-Clause license. Note that NO WARRANTY is provided.
  * See "LICENSE_BSD2.txt" for details.
  *
- * @TAG(DATA61_BSD)
+ * @TAG(NICTA_BSD)
  */
 
 #include "browser.h"
@@ -18,11 +16,12 @@
 #include "regions.h"
 #include "printer.h"
 
+
 #include <stdio.h>
 #include <string.h>
 
 static int
-_browse_tables(void* table, size_t offset)
+_browse_tables(void* table, uint32_t offset)
 {
     /*
      * NOTE: offset added to address JUST BEFORE moving
@@ -45,7 +44,7 @@ _browse_tables(void* table, size_t offset)
             acpi_print_table(table);
             acpi_rsdp_t* rsdp = (acpi_rsdp_t*)table;
             vector[0] = (void*)rsdp->rsdt_address;
-            vector[1] = (void*)(uintptr_t)rsdp->xsdt_address;
+            vector[1] = (void*)(unsigned int)rsdp->xsdt_address;
             printf("\n");
             printf("a - RSDT at %p\n", vector[0]);
             printf("b - XSDT at %p\n", vector[1]);
@@ -79,7 +78,7 @@ _browse_tables(void* table, size_t offset)
             while (entry != NULL) {
                 char sig[5];
                 sig[4] = '\0';
-                char* _e = (char*)(uintptr_t)(*entry);
+                char* _e = (char*)(uint32_t)(*entry);
                 memcpy(sig, _e, 4);
                 vector[i] = _e;
                 printf("%c - %s table at %p\n", i + 'a', sig, vector[i]);
@@ -94,8 +93,8 @@ _browse_tables(void* table, size_t offset)
             acpi_fadt_t* fadt = (acpi_fadt_t*)table;
             vector[0] = (void*)fadt->facs_address;
             vector[1] = (void*)fadt->dsdt_address;
-            vector[2] = (void*)(uintptr_t)fadt->x_facs_address;
-            vector[3] = (void*)(uintptr_t)fadt->x_dsdt_address;
+            vector[2] = (void*)(unsigned int)fadt->x_facs_address;
+            vector[3] = (void*)(unsigned int)fadt->x_dsdt_address;
             printf("\n");
             printf("a -  FACS at %p\n", vector[0]);
             printf("b -  DSDT at %p\n", vector[1]);
@@ -108,7 +107,7 @@ _browse_tables(void* table, size_t offset)
             acpi_print_table(table);
             acpi_facs_t* facs = (acpi_facs_t*)table;
             vector[0] = (void*)facs->firmware_walking_vector;
-            vector[3] = (void*)(uintptr_t)facs->x_firmware_walking_vector;
+            vector[3] = (void*)(unsigned int)facs->x_firmware_walking_vector;
             printf("\n");
             printf("a - Firware Walking vector at %p\n", vector[0]);
             printf("b - X Firware walking vector at %p\n", vector[1]);
@@ -219,10 +218,11 @@ _browse_tables(void* table, size_t offset)
 }
 
 void
-acpi_browse_tables(const acpi_rsdp_t* rsdp, size_t offset)
+acpi_browse_tables(const acpi_rsdp_t* rsdp, uint32_t offset)
 {
     _browse_tables((void*)rsdp, offset);
 }
+
 
 static int
 _browse_regions(const RegionList_t* rlist, int parent)
@@ -292,6 +292,7 @@ _browse_regions(const RegionList_t* rlist, int parent)
     }
     while (1);
 }
+
 
 void
 acpi_browse_regions(const RegionList_t* rlist)
